+++ integrations/opensearch/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [integrations/opensearch-v6.1.2] - 2026-01-16
+
+### 🐛 Bug Fixes
+
+- Docs: fixing docstring in `OpenSearchDocumentStore` (#2758)
+
+
+## [integrations/opensearch-v6.1.1] - 2026-01-16
+
+### 🐛 Bug Fixes
+
+- Docs: fixing docstrings in `OpenSearchDocumentStore` (#2757)
+
+
+## [integrations/opensearch-v6.1.0] - 2026-01-16
+
+### 🚀 Features
+
+- Adding  count with filtering operations to `OpenSearchDocumentStore` (#2653)
+
+
 ## [integrations/opensearch-v6.0.0] - 2026-01-13
 
 ### 🧪 Testing